finish category screen ui
